Relocate dextre command
