--- mss/dextre/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/poka-yoke/spaceflight/mss/dextre/cmd"
-
-func main() {
-	cmd.Execute()
-}
