+++ pkg/sidekiq/sidekiq_test.go
